refactor(security): consolidate security endpoints
Moved addPolicy and hasPermission endpoints from CasbinController to SecurityController
Removed getRoleNames method from User class
Streamlined security service integration and endpoint organization

This change consolidates security-related endpoints into a single controller, removes deprecated user role methods, and improves overall security endpoint organization while maintaining existing functionality.
