+++ src/main/java/org/endeavourhealth/imapi/model/security/User.kt
@@ -32,9 +32,4 @@ class User {
   fun adminSetPassword(password: String) {
     this.password = password
   }
-
-  // required for casbin
-  fun getRoleNames(): List<String> {
-    return this.roles.map { it.name }
-  }
 }
